feat: add service payment
