+++ app/Http/Controllers/PaymentController.php
@@ -3,35 +3,28 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\PaymentPostRequest;
-use App\Jobs\ProcessPayment;
-use App\Models\Payment;
-use App\Services\Payments\PaymentProcessorFactory;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Log;
+use App\Jobs\ProcessPaymentJob;
+use App\Services\Contracts\PaymentServiceInterface;
 
 class PaymentController extends Controller
 {
 
+    protected $paymentService;
+
+    public function __construct(PaymentServiceInterface $paymentService)
+    {
+        $this->paymentService = $paymentService;
+    }
+
     public function store(PaymentPostRequest $request)
     {
         $validated = $request->validated();
-        ProcessPayment::dispatch($validated);
+        ProcessPaymentJob::dispatch($validated, $this->paymentService);
+        //$this->paymentService->processPayment($validated);
         return response()->json([
             "status" => "processing",
             "message" => "Payment request received and is being processed.",            
         ], 202);
     }
 
-    public function processPayment(array $validated)
-    {
-        try {
-            $paymentMethod = $validated['payment_method'];
-            $processor = PaymentProcessorFactory::create($paymentMethod);
-            $payment = Payment::create($validated);
-            $processor->process($payment);            
-            Log::info('Payment processed successfully: ' . $payment->id);
-        } catch (\Exception $e) {
-            Log::error('Error processing payment: ' . $e->getMessage());
-        }
-    }
 }
