+++ app/Services/Contracts/PaymentServiceInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Services\Contracts;
+
+interface PaymentServiceInterface
+{
+    public function processPayment(array $data): void;
+}
