+++ app/Services/Payments/BoletoProcessorPaymentService.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace App\Services\Contracts\Payments;
+namespace App\Services\Payments;
 
 use App\Models\Payment;
 use App\Services\Payments\Contracts\PaymentProcessorInterface;
 use Exception;
 
-class CreditCardProcessorPaymentService implements PaymentProcessorInterface
+class BoletoProcessorPaymentService implements PaymentProcessorInterface
 {
     public function process(Payment $payment): void
     {
         try {
             $payment->status = 'processed';
             $payment->save();
-        } catch (Exception $e) {
+        } catch (\Exception $e) {
             $payment->status = 'failed';
             $payment->save();
             throw new Exception("processed failed:" . $e->getMessage());
