+++ app/Services/Payments/PixProcessorPaymentService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Contracts\Payments;
+namespace App\Services\Payments;
 
 use App\Models\Payment;
 use App\Services\Payments\Contracts\PaymentProcessorInterface;
